refactor: Extract Meshtastic handler from rns_bridge.py
- Create new MeshtasticHandler class (595 lines) with dependency injection
- Reduce rns_bridge.py from 1,991 to 1,587 lines (~20% reduction)
- Handler encapsulates connection, messaging, and reconnection logic
- Uses callback pattern for message/status notifications to bridge
- Maintains thread-safety with shared stats lock

Extracted methods:
- connect/disconnect for connection management
- send_text/queue_send for message sending
- run_loop for main connection loop with reconnect
- _on_receive/_handle_text_message for message processing
- _discover_relay_node for Meshtastic 2.6+ relay tracking
- _poll/_handle_connection_lost for health monitoring

Issue #6: File size refactoring
